pyproject.toml
[project]
name = "e_memory"
version = "0.1.0"
description = ""
authors = [
    {name = "Lucas Rocha",email = "lucash.rocha@hotmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "django (>=5.2,<6.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "mysqlclient (>=2.2.7,<3.0.0)",
    "pymongo (>=4.12.0,<5.0.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "djoser (>=2.3.1,<3.0.0)"
]

[tool.poetry]
package-mode = false
packages = [{include = "e_memory", from = "src"}]

[tool.poetry.scripts]
start = "uvicorn e_memory.main:app"
dev = "uvicorn --reload e_memory.main:app"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.1.3"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
